// src/types.d.ts
export interface Image {
  src: string;
  alt?: string;
}
// ...
export interface Contact {
  name: string;
  title: string;
  email: string;
  bio: string;
  img: Image;
}